\name{sim_dat}
\docType{data}
\alias{sim_dat}
\title{A simulated multivariate data}
\description{
A simulated multivariate data with 30 samples on 15 features.
}
\section{Details}{
The data set contains a 30 by 15 matrix, with mean of each feature being -4/15, -3/15 ... to 10/15. Covariance among
different features being 0.1 and the variance for each is 1. This is the complete data without any missingness.
}
\keyword{datasets}
